refactor tests and create a shared test suite
